Huge Commit - Readme.pls:
Bug:
The bug resulted in showing a delay after the player busted.
Game status was displayed after the delay, instead of
immediately.

Bug Fix:
Game result is shown now immediately after the player busts.

In the process of fixing the bug, I ended up changing and (hopefully)
simplifying our codebase. Every change I made resulted in another change,
leading to A LOT of changes and I had a hard time, splitting them up into
smaller commits, hence this widespread commit.

At the end of the game, the game result is displayed as a status message.

Also the game now greys out the area of the loser, which highlights the
area of the winner indirectly.

Todo: add dialog, for starting a new game or exiting the game.